day_1_ddo/ddo_hackathon/requirements.txt
cobyqa==1.1.1
matplotlib==3.9.2
numpy==2.1.3
pandas==2.2.3
PyGithub==2.5.0
python-dotenv==1.0.1
scipy==1.14.1
tqdm==4.67.0
google-cloud-storage
